rename: Daily -> Hourly, Weekly -> Daily로 수정
